Tests/SetSniffPropertyTest: rename fixtures
With an eye on adding additional tests for the Ruleset class, which would require additional test fixtures, this commit renames the `Category` subdirectory to `SetProperty` to directly link the fixtures in that folder to the `SetSniffPropertyTest`.

It also removes the `SetProperty` prefix from the dummy sniffs in that folder as it is not necessary anymore what with the folder being renamed.

Includes making the necessary updates to the dummy sniffs and the tests to allow for the rename.
